Memo line component
