+++ packs/step-parser/src/index.ts
@@ -1,7 +1,3 @@
 import { parseSteps } from "step-parser";
 
-export default parseSteps;
-
-export function flat<T>(nestedArrays: T[][]) {
-  return nestedArrays.flat();
-}
+export { parseSteps };
